Challenge deploy

+++ lib/deployment/deployment.go
@@ -5,11 +5,14 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"log"
 	"path/filepath"
 	"text/template"
 
 	g "github.com/sdslabs/katana/configs"
 	"github.com/sdslabs/katana/lib/utils"
+	appsv1 "k8s.io/api/apps/v1"
+	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -71,6 +74,11 @@ func ApplyManifest(kubeconfig *rest.Config, kubeclientset *kubernetes.Clientset,
 
 		if _, err := dri.Create(context.Background(), unstructuredObj, metav1.CreateOptions{}); err != nil {
 			if _, err := dri.Update(context.Background(), unstructuredObj, metav1.UpdateOptions{}); err != nil {
+				if unstructuredObj.GetObjectKind().GroupVersionKind().Kind == "PersistentVolumeClaim" {
+					// Skip PVCs
+					continue
+					// TODO: Handle PVCs, currently on deletion of PVCs, the cluster is stuck in a loop
+				}
 				_ = dri.Delete(context.Background(), unstructuredObj.GetName(), metav1.DeleteOptions{})
 				watcher, err := dri.Watch(context.Background(), metav1.ListOptions{
 					FieldSelector: fmt.Sprintf("metadata.name=%s", unstructuredObj.GetName()),
@@ -105,10 +113,16 @@ func DeployCluster(kubeconfig *rest.Config, kubeclientset *kubernetes.Clientset)
 
 	deploymentConfig := utils.DeploymentConfig()
 
-	for _, m := range clusterConfig.Manifests {
+	clientset, _ := utils.GetKubeClient()
+
+	nodes, _ := utils.GetNodes(clientset)
+
+	deploymentConfig.NodeAffinityValue = nodes[0].Name
+
+	for _, m := range clusterConfig.TemplatedManifests {
 		manifest := &bytes.Buffer{}
-		fmt.Printf("Applying: %s\n", m)
-		tmpl, err := template.ParseFiles(filepath.Join(clusterConfig.ManifestDir, m))
+		log.Printf("Applying: %s\n", m)
+		tmpl, err := template.ParseFiles(filepath.Join(clusterConfig.TemplatedManifestDir, m))
 		if err != nil {
 			return err
 		}
@@ -120,5 +134,82 @@ func DeployCluster(kubeconfig *rest.Config, kubeclientset *kubernetes.Clientset)
 		}
 	}
 
+	return nil
+}
+
+func DeployChallengeToCluster(challengeName, teamName string, firstPatch bool, replicas int32) error {
+
+	teamNamespace := teamName + "-ns"
+	kubeclient, _ := utils.GetKubeClient()
+
+	deploymentsClient := kubeclient.AppsV1().Deployments(teamNamespace)
+	imageName := "harbor.katana.local/katana/" + challengeName + ":latest"
+	if firstPatch {
+		/// Retrieve the existing deployment
+		existingDeployment, err := deploymentsClient.Get(context.TODO(), challengeName, metav1.GetOptions{})
+		if err != nil {
+			log.Println("Error in retrieving existing deployment.")
+			log.Println(err)
+			return err
+		}
+
+		existingDeployment.Spec.Template.Spec.Containers[0].Image = "harbor.katana.local/katana/" + teamName + "-" + challengeName + ":latest"
+
+		_, err = deploymentsClient.Update(context.TODO(), existingDeployment, metav1.UpdateOptions{})
+		if err != nil {
+			log.Println("Error in updating deployment.")
+			log.Println(err)
+			return err
+		}
+
+		log.Println("Updated deployment with new image.")
+		return nil
+	}
+
+	manifest := &appsv1.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: teamNamespace,
+			Name:      challengeName,
+		},
+		Spec: appsv1.DeploymentSpec{
+			Replicas: &replicas,
+			Selector: &metav1.LabelSelector{
+				MatchLabels: map[string]string{
+					"app": challengeName,
+				},
+			},
+			Template: v1.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Labels: map[string]string{
+						"app": challengeName,
+					},
+				},
+				Spec: v1.PodSpec{
+					Containers: []v1.Container{
+						{
+							Name:            challengeName + "-" + teamName,
+							Image:           imageName,
+							ImagePullPolicy: v1.PullPolicy("IfNotPresent"),
+							Ports: []v1.ContainerPort{
+								{
+									Name:          "challenge-port",
+									ContainerPort: 3000,
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+	log.Println("Creating deployment...")
+	result, err := deploymentsClient.Create(context.TODO(), manifest, metav1.CreateOptions{})
+
+	if err != nil {
+		log.Println("Unable to create deployement")
+		panic(err)
+	}
+
+	log.Printf("Created deployment %q.\n", result.GetObjectMeta().GetName()+" in namespace "+teamNamespace)
 	return nil
 }
